requirements.txt
setuptools
-e .
fastapi
uvicorn
langchain
langchain-openai
langchain_community
langchain_core
langchain_pinecone
python-dotenv
faiss-cpu
pypdf
# pymongo
# pinecone-client
# langchain-mongodb